+++ .stylelintrc.cjs
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: ['./tools/configs/.stylelintrc-base.cjs']
+}