+++ apps/api/src/Z_COMMON/api-bio-types/species.ts
@@ -1,0 +1,31 @@
+import { ExtinctionRiskCode } from '../iucn'
+
+export const SPECIES_SOURCE_IUCN = 'IUCN'
+export const SPECIES_SOURCE_WIKI = 'Wiki'
+
+const SPECIES_SOURCES = <const>[SPECIES_SOURCE_IUCN, SPECIES_SOURCE_WIKI]
+export type SpeciesSource = typeof SPECIES_SOURCES[number]
+
+export interface SpeciesInformation {
+  description: string
+  sourceUrl: string | undefined
+  sourceType: SpeciesSource
+}
+
+export interface SpeciesExternalLink {
+  title: string
+  sourceUrl: string | undefined
+  sourceType: SpeciesSource
+}
+
+export interface Species {
+  speciesId: number
+  speciesSlug: string
+  scientificName: string
+  extinctionRisk: ExtinctionRiskCode
+  thumbnailImageUrl?: string
+  taxonId: number
+  taxon: string
+  information?: SpeciesInformation | undefined
+  externalLinks?: SpeciesExternalLink[]
+}