+++ apps/api/src/dashboard/dashboard-generated-controller.ts
@@ -1,0 +1,20 @@
+import { Controller } from '../_services/api-helper/types.js'
+import { assertParamsExist } from '../_services/validation/index.js'
+import { DashboardGeneratedParams, DashboardGeneratedResponse } from '../Z_COMMON/api-bio-types/dashboard-generated.js'
+import { getGeneratedData } from './dao.js'
+
+export const dashboardGeneratedController: Controller<DashboardGeneratedParams, DashboardGeneratedResponse> = async (req) => {
+  // Inputs & validation
+  const { projectId } = req.params
+  assertParamsExist({ projectId })
+
+  // Query
+  const data = await getGeneratedData()
+
+  // Response
+  const response: DashboardGeneratedResponse = {
+    ...data
+  }
+
+  return response
+}