+++ apps/website/src/dashboard/components/dashboard-highlighted-species/dashboard-highlighted-species.ts
@@ -1,0 +1,25 @@
+import { Vue } from 'vue-class-component'
+import { Inject, Prop } from 'vue-property-decorator'
+
+import { ExtinctionRisk } from '@rfcx-bio/common/iucn'
+
+import { RouteNames } from '~/router'
+import { BiodiversityStore } from '~/store'
+
+export interface HighlightedSpeciesRow {
+  speciesId: number
+  speciesName: string
+  speciesSlug: string
+  imageUrl: string
+  extinctionRisk: ExtinctionRisk
+}
+
+export default class DashboardHighlightedSpecies extends Vue {
+  @Inject() readonly store!: BiodiversityStore
+  @Inject() readonly ROUTE_NAMES!: RouteNames
+  @Prop() species!: HighlightedSpeciesRow[]
+
+  get hasData (): boolean {
+    return this.species.length > 0
+  }
+}