+++ apps/website/src/dashboard/components/dashboard-line-chart/dashboard-line-chart.vue
@@ -1,0 +1,8 @@
+<template>
+  <line-chart-component
+    dom-id="dashboard-line-chart"
+    :config="config"
+    :datasets="chartData"
+  />
+</template>
+<script lang="ts" src="./dashboard-line-chart.ts"></script>