+++ packages/common/src/index.ts
@@ -1,1 +1,5 @@
+import { urlify } from '@rfcx-bio/utils/url-helpers'
+
 console.info('Mock `common` module')
+
+console.log(urlify('potato'))