+++ packages/utils/jest.config.js
@@ -1,11 +1,4 @@
-export default {
-  rootDir: 'src',
-  moduleFileExtensions: [
-    'js',
-    'ts',
-    'json'
-  ],
-  transform: {
-    '^.+\\.ts$': 'ts-jest'
-  }
-}
+import { createConfig } from '../../tools/configs/jest-config-factory.js'
+import tsconfig from './tsconfig.build.json' // Must be the tsconfig that defines `paths`
+
+export default createConfig(tsconfig)